Switch to using injectable to setup route for getting current port forward
