+++ src/main/routes/index.ts
@@ -5,5 +5,4 @@
 
 export * from "./kubeconfig-route";
 export * from "./metrics-route";
-export * from "./port-forward-route";
 export * from "./version-route";
